requirements.txt
numpy
gym==0.9.5
shapely
image
pygame>=1.9.3
scipy
ray
scikit-image
opencv-python
codecov